Allow Issue to get deleted Magazine and Mag_Issue
The `paranoia` gem adds a `default_scope` to `Spree::Product` that hides all deleted Products.

However, I can imagine that a site admin might want to see all the issues in a subscription even if those products are old and have been deleted.

This syntax allows the getter methods to use the unscoped versions of the Product model.

Note that this applies to `Issue.first.magazine` but does _not_ apply to `Issue.includes(:magazine).first.magazine` -- this is a known and outstanding issue in the `paranoia` gem.