Don't throw exception when product is soft-deleted
A long-running store may have Issue objects attached to Products which have been soft-deleted and removed from sale. For bookkeeping reasons, account history, and the like, we want to still be able to allow Issues to know which Magazine they're issues of. This is accomplished simply by using the `with_deleted` scope.