Cleaner syntax to get Subscriptions from Issues
We can use ActiveRecord here to avoid the n+1 queries problem where the earlier syntax would call all of the ShippedIssue objects one at a time from the db.  It also has the advantage of returning an ActiveRecord relation rather than an array, so it's extensible by other scopes, etc.